okta-vite-fastapi/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
python-jose[cryptography]==3.3.0
httpx==0.27.0
authlib==1.3.0
